1695. Maximum Erasure Value
